requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
distro==1.9.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
numpy==2.4.4
pandas==3.0.2
pybase64==1.4.3
pydantic==2.13.4
pydantic_core==2.46.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
